+++ source/Attenuator/Colours.h
@@ -59,7 +59,7 @@ uint8_t getBrightness(uint8_t i_percent = 100) {
 // This must match the number of device ENUM entries (though that is rarely changed).
 uint8_t i_curr_colour[DEVICE_NUM_LEDS] = { 0, 0, 0 };
 uint8_t i_curr_bright[DEVICE_NUM_LEDS] = { 0, 0, 0 };
-int i_next_bright[DEVICE_NUM_LEDS] = { -1, -1, -1 }; // Uses int to allow negative steps.
+int16_t i_next_bright[DEVICE_NUM_LEDS] = { -1, -1, -1 }; // Uses int to allow negative steps.
 uint8_t i_count[DEVICE_NUM_LEDS] = { 0, 0, 0 }; // Counter-based changes for certain themes.
 millisDelay ms_color_change[DEVICE_NUM_LEDS]; // Timers for changing colors for certain themes.
 uint16_t i_change_delay[DEVICE_NUM_LEDS] = { 10, 10, 10 }; // Default delay time for changes.
